conditionally load beans
