+++ src/main/java/io/numaproj/kafka/KafkaApplication.java
@@ -1,21 +1,38 @@
 package io.numaproj.kafka;
 
-import io.numaproj.numaflow.sourcer.Server;
+import io.numaproj.kafka.consumer.Admin;
+import io.numaproj.kafka.consumer.KafkaSourcer;
+import io.numaproj.kafka.consumer.Worker;
+import io.numaproj.kafka.producer.KafkaSinker;
+import io.numaproj.kafka.schema.ConfluentRegistry;
+import io.numaproj.kafka.schema.Registry;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.builder.SpringApplicationBuilder;
+import org.springframework.boot.context.properties.EnableConfigurationProperties;
+import org.springframework.context.annotation.ComponentScan;
+import org.springframework.context.annotation.FilterType;
 
 @Slf4j
 @SpringBootApplication
+@EnableConfigurationProperties
+@ComponentScan(
+    excludeFilters = {
+      @ComponentScan.Filter(
+          type = FilterType.ASSIGNABLE_TYPE,
+          classes = {
+            KafkaSinker.class,
+            KafkaSourcer.class,
+            Admin.class,
+            Worker.class,
+            ConfluentRegistry.class,
+            Registry.class
+          })
+    })
 public class KafkaApplication {
 
-    public static void main(String[] args) throws Exception {
-        log.info("Supplied arguments:{}", (Object) args);
-        var ctx = new SpringApplicationBuilder(KafkaApplication.class)
-                .run(args);
-        // TODO - conditionally load beans based on the numaflow.handler property
-        Server server = ctx.getBean(Server.class);
-        log.info("Starting the Kafka source application...");
-        server.start();
-    }
+  public static void main(String[] args) {
+    log.info("Supplied arguments: {}", (Object) args);
+    new SpringApplicationBuilder(KafkaApplication.class).run(args);
+  }
 }
